Fix printf-like format strings everywhere
There were various bugs in using formatting strings:
 * Variable strings passed as format strings (see issue 906 as an example)
 * Wrong data size for integer types (%d for size_t)
 * Plain format errors (%s to print out character)
 * Typos ("Setting output directory to s", OUTPUTDIR)
 * Unused formatting parameters

+++ src/mon_processes.c
@@ -56,7 +56,7 @@ RawSaveItemList(userList,vbuff);
 
 DeleteItemList(userList);
 
-CfOut(cf_verbose,"","(Users,root,other) = (%d,%d,%d)\n",cf_this[ob_users],cf_this[ob_rootprocs],cf_this[ob_otherprocs]);
+CfOut(cf_verbose,"","(Users,root,other) = (%d,%d,%d)\n",(int)cf_this[ob_users],(int)cf_this[ob_rootprocs],(int)cf_this[ob_otherprocs]);
 }
 
 static int GatherProcessUsers(struct Item **userList, int *userListSz, int *numRootProcs, int *numOtherProcs)
